fix: no cyclic imports
this is why I didn't do this earlier...
