+++ spras/config/algorithms.py
@@ -5,7 +5,7 @@ and rather mainly contains validators and lower-level pydantic code.
 """
 import ast
 import copy
-from typing import Annotated, Any, Callable, Literal, Optional, Union, cast, get_args
+from typing import Annotated, Any, Callable, Literal, Union, cast, get_args
 
 import numpy as np
 from pydantic import (
@@ -17,27 +17,11 @@ from pydantic import (
     create_model,
 )
 
+from spras.config.runs import RunSettings
 from spras.runner import algorithms
 
 # This contains the dynamically generated algorithm schema for use in `schema.py`
 __all__ = ['AlgorithmUnion']
-
-def validate_duration(value):
-    parsed_duration = value(value, granularity='seconds')
-    if not parsed_duration: raise RuntimeError(f"Encountered unparsable duration string '{value}'.")
-
-PyDateTimeDuration = Annotated[
-    int,
-    BeforeValidator(validate_duration)
-]
-
-class RunSettings(BaseModel):
-    """All of the non-parameter settings associated with a run."""
-
-    timeout: Optional[PyDateTimeDuration] = None
-    """The associated timeout with a run, parsed with `pytimeparse`."""
-
-    model_config = ConfigDict(extra='forbid', use_attribute_docstrings=True)
 
 def is_numpy_friendly(type: type[Any] | None) -> bool:
     """
