+++ spras/domino.py
@@ -5,8 +5,8 @@ from typing import Optional
 import pandas as pd
 from pydantic import BaseModel, ConfigDict
 
-from spras.config.algorithms import RunSettings
 from spras.config.container_schema import ProcessedContainerSettings
+from spras.config.runs import RunSettings
 from spras.config.util import BaseModel
 from spras.containers import ContainerError, prepare_volume, run_container_and_log
 from spras.interactome import (
