+++ spras/responsenet.py
@@ -2,8 +2,8 @@ from pathlib import Path
 
 from pydantic import BaseModel, ConfigDict
 
-from spras.config.algorithms import RunSettings
 from spras.config.container_schema import ProcessedContainerSettings
+from spras.config.runs import RunSettings
 from spras.containers import prepare_volume, run_container_and_log
 from spras.interactome import (
     convert_undirected_to_directed,
